+++ wgpu/src/backend/webgpu.rs
@@ -1563,6 +1563,40 @@ impl dispatch::InstanceInterface for ContextWebGpu {
         // Devices are automatically polled.
         true
     }
+
+    #[cfg(feature = "wgsl")]
+    fn wgsl_language_features(&self) -> crate::WgslLanguageFeatures {
+        let mut wgsl_language_features = crate::WgslLanguageFeatures::empty();
+        if let Some(gpu) = &self.gpu {
+            gpu.wgsl_language_features()
+                .keys()
+                .into_iter()
+                .map(|wlf| wlf.expect("`WgslLanguageFeatures` elements should be valid"))
+                .map(|wlf| {
+                    wlf.as_string()
+                        .expect("`WgslLanguageFeatures` should be string set")
+                })
+                .filter_map(|wlf| match wlf.as_str() {
+                    "readonly_and_readwrite_storage_textures" => {
+                        Some(crate::WgslLanguageFeatures::ReadOnlyAndReadWriteStorageTextures)
+                    }
+                    "packed_4x8_integer_dot_product" => {
+                        Some(crate::WgslLanguageFeatures::Packed4x8IntegerDotProduct)
+                    }
+                    "unrestricted_pointer_parameters" => {
+                        Some(crate::WgslLanguageFeatures::UnrestrictedPointerParameters)
+                    }
+                    "pointer_composite_access" => {
+                        Some(crate::WgslLanguageFeatures::PointerCompositeAccess)
+                    }
+                    _ => None,
+                })
+                .for_each(|wlf| {
+                    wgsl_language_features |= wlf;
+                })
+        }
+        wgsl_language_features
+    }
 }
 
 impl Drop for ContextWebGpu {